# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/security/authorize": {
      "post": {
        "operationId": "AuthorizationController_authorize",
        "summary": "Paso 1: valida credenciales y emite un código de autorización",
        "description": "El cliente genera un `code_verifier` aleatorio y envía su hash S256 como `code_challenge`. Devuelve una redirección por cada sistema al que el usuario tiene acceso; todas llevan el mismo código.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Código emitido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Login exitoso"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string",
                            "example": "/validate?code=abc&cid=xyz"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cuerpo inválido (campo obligatorio ausente o typeDevice desconocido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Usuario o contraseña incorrectos (mismo error si el usuario no existe).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "El correo aún no está verificado. El cuerpo NO usa el sobre estándar: `{ code: \"EMAIL_NOT_VERIFIED\", email }`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Email no verificado"
                    },
                    "data": {
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Autorización (Authorization Code + PKCE)"
        ]
      }
    },
    "/security/token": {
      "post": {
        "operationId": "AuthorizationController_token",
        "summary": "Paso 2: canjea el código (+ code_verifier) por la sesión",
        "description": "Verifica PKCE y el tipo de dispositivo, consume el código (un solo uso), marca la sesión como autenticada (cookie `auth.session`) y fija la cookie HttpOnly `auth.refresh`.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión establecida.",
            "headers": {
              "Set-Cookie": {
                "description": "auth.refresh (HttpOnly, SameSite=Lax) y auth.session.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Callback exitoso"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Autenticación exitosa"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Código inexistente/consumido, verifier PKCE inválido o dispositivo distinto.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Autorización (Authorization Code + PKCE)"
        ]
      }
    },
...
    "/security/password-reset/request": {
      "post": {
        "operationId": "PasswordResetController_request",
        "summary": "Solicita el enlace de restablecimiento",
        "description": "Responde siempre el mismo mensaje genérico (exista o no el correo, esté o no activa la cuenta). Estado actual: el envío del correo aún NO está implementado.",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud aceptada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Si el correo existe, recibirás un enlace de restablecimiento"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Correo con formato inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Recuperación de contraseña"
        ]
      }
    },
    "/security/password-reset/validate": {
      "get": {
        "operationId": "PasswordResetController_validate",
        "summary": "Comprueba si el token del enlace sigue siendo válido",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "Token del enlace de restablecimiento.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uuid",
            "required": true,
            "in": "query",
            "description": "Uuid del token de restablecimiento (parámetro `uuid` del enlace).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado de la validación.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "email": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Recuperación de contraseña"
        ]
      }
    },
    "/security/password-reset/reset": {
      "post": {
        "operationId": "PasswordResetController_reset",
        "summary": "Establece la nueva contraseña con el token del enlace",
        "description": "Marca el token como usado y revoca las sesiones del usuario en base de datos.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña restablecida.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contraseña restablecida exitosamente"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Contraseñas distintas, token inválido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Recuperación de contraseña"
        ]
      }
    },
...
  },
  "info": {
    "title": "ms-identity",
    "description": "Identidad y autenticación de SEIS App: Authorization Code + PKCE, ciclo de sesión, recuperación de contraseña, registro y perfil de usuario.\n\n**Flujo de login**: `POST /security/authorize` (credenciales + code_challenge) → código → `POST /security/token` (código + code_verifier) → cookies `auth.session` y `auth.refresh`. Las rutas de perfil usan el JWT de acceso como Bearer.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth.session"
      },
      "refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth.refresh"
      }
    },
    "schemas": {
      "AuthorizeRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "sparra"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "code_challenge": {
            "type": "string",
            "description": "PKCE: BASE64URL(SHA256(code_verifier)), método S256."
          },
          "typeDevice": {
            "type": "string",
            "enum": [
              "WEB",
              "DESKTOP",
              "MOBILE",
              "POSTMAN"
            ]
          },
          "CorrelationId": {
            "type": "string",
            "description": "Id de correlación de este intento; debe repetirse como `cid` en /security/token."
          }
        },
        "required": [
          "CorrelationId",
          "code_challenge",
          "password",
          "typeDevice",
          "username"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Usuario o contraseña incorrectos"
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
      "TokenRequestDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código de autorización devuelto por /security/authorize (un solo uso)."
          },
          "typeDevice": {
            "type": "string",
            "enum": [
              "WEB",
              "DESKTOP",
              "MOBILE",
              "POSTMAN"
            ],
            "description": "Debe coincidir con el usado en /security/authorize."
          },
          "codeVerifier": {
            "type": "string",
            "description": "PKCE: el code_verifier original cuyo hash se envió como code_challenge."
          },
          "cid": {
            "type": "string",
            "description": "CorrelationId enviado en /security/authorize."
          }
        },
        "required": [
          "code",
          "typeDevice",
          "codeVerifier",
          "cid"
        ]
      },
...
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "correo": {
            "type": "string",
            "example": "ana@correo.cl"
          }
        },
        "required": [
          "correo"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "format": "password"
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 8,
            "format": "password",
            "description": "Debe ser igual a newPassword."
          }
        },
        "required": [
          "token",
          "uuid",
          "newPassword",
          "confirmPassword"
        ]
      },
...
    }
  }
}